services/api/requirements.txt
fastapi
uvicorn
boto3
python-multipart
pydantic
pyjwt
